Add files via upload
Uploaded script to encrypt/decrypt file based on inputs

Project: Assignment: Publish a Python Module on GitHub

Topic:  Text encryption/decryption

Topic Description:
Create a command line interface where you can supply a message and a key value to encrypt a message using the python cryptography module. The message will be written to a file of the user's choice. The script must be able to read text from a specified file and be able to decrypt the message using a supplied key value.